test: remove retired eager SpD diagnostics
